Test format_test_case
